requirements.txt
boto3
python3-picamzero
